+++ modules/acf/js/index.js
@@ -1,3 +1,3 @@
-import './common';
-import './fields';
+import './register';
+import './switch';
 import './qtranslatex';
